RVL SDK: Rename DC to OSCache in accordance with OpenRVL
