+++ source/dol/bases/d_res_info.cpp
@@ -1,7 +1,7 @@
 #include <dol/bases/d_res.hpp>
 #include <dol/mLib/m_heap.hpp>
 #include <lib/rvl/vi/VI.h>
-#include <lib/rvl/dc/DC.h>
+#include <lib/rvl/os/OSCache.h>
 #include <lib/MSL_C/string.h>
 
 dRes_c::info_c::info_c() :
